Implement pricing strategies and availability observer

+++ src/main/java/org/example/vehicle/Bike.java
@@ -5,10 +5,10 @@ import java.util.Objects;
 public class Bike implements   Vehicle {
     private final String id;
     private final String model;
-    private final int basePricePerDay;
-    private final int basePricePerHour;
+    private final float basePricePerDay;
+    private final float basePricePerHour;
 
-    public Bike(String id, String model, int basePricePerDay, int basePricePerHour) {
+    public Bike(String id, String model, float basePricePerDay, float basePricePerHour) {
         this.id = Objects.requireNonNull(id);
         this.model = Objects.requireNonNull(model);
         this.basePricePerDay = basePricePerDay;
@@ -18,8 +18,8 @@ public class Bike implements   Vehicle {
     @Override public String getId() { return id; }
     @Override public String getType() { return "Bike"; }
     @Override public String getModel() { return model; }
-    @Override public int getBasePricePerDay() { return basePricePerDay; }
-    @Override public int getBasePricePerHour() { return basePricePerHour; }
+    @Override public float getBasePricePerDay() { return basePricePerDay; }
+    @Override public float getBasePricePerHour() { return basePricePerHour; }
     @Override public String getDescription() {
         return String.format("Bike[id=%s, model=%s]", id, model);
     }
